Fix param length
