+++ packages/docs/pages/docs/library/styles.mdx
@@ -180,7 +180,7 @@ control sequences that would otherwise be used to render text in a terminal.
 
 #### Indentation level
 
-A terminal string optionally has a starting column: the indentation level respective to the first
+A terminal string optionally has an indentation level: the starting column respective to the first
 terminal column. When wrapping text to a desired width, it tries to respect this setting, but if
 the desired width does not accommodate the text while respecting its indentation level, the text is
 instead wrapped relative to the beginning of each line.
