+++ packages/tsargp/test/formatter/formatter.formatting.spec.ts
@@ -205,19 +205,15 @@ describe('HelpFormatter', () => {
         flag1: {
           type: 'flag',
           names: ['-f', null, '--flag'],
-          desc: 'A flag option',
         },
         flag2: {
           type: 'flag',
           names: [null, '--flag2', null],
-          desc: 'A flag option',
         },
       } as const satisfies Options;
       const config: HelpConfig = { align: { names: 'left' } };
       const message = new HelpFormatter(new OptionValidator(options), config).formatHelp();
-      expect(message.wrap()).toEqual(
-        '  -f, --flag    A flag option\n  --flag2       A flag option\n',
-      );
+      expect(message.wrap()).toEqual('  -f, --flag\n  --flag2\n');
     });
 
     it('should align option names to the right boundary', () => {
@@ -225,19 +221,15 @@ describe('HelpFormatter', () => {
         flag1: {
           type: 'flag',
           names: ['-f', null, '--flag'],
-          desc: 'A flag option',
         },
         flag2: {
           type: 'flag',
           names: [null, '--flag2', null],
-          desc: 'A flag option',
         },
       } as const satisfies Options;
       const config: HelpConfig = { align: { names: 'right' } };
       const message = new HelpFormatter(new OptionValidator(options), config).formatHelp();
-      expect(message.wrap()).toEqual(
-        '  -f, --flag    A flag option\n     --flag2    A flag option\n',
-      );
+      expect(message.wrap()).toEqual('  -f, --flag\n     --flag2\n');
     });
   });
 });
